requirements.txt
requests
python-dotenv
mssql_python